requirements.txt
flask==3.0.3
argon2-cffi==23.1.0
python-dotenv==1.0.1
